re structure backend code. now the core logics sits in 'core' and the services will have it's own service logic or APIs accordingly.
