+++ core/data_models/models.py
@@ -1,12 +1,10 @@
-# generated by fastapi-codegen:
-#   filename:  /Users/sanjay_acorn/Downloads/openapi.yaml
-#   timestamp: 2023-11-18T07:40:51+00:00
-
 from __future__ import annotations
 
+# Standard Library
 from enum import Enum
 from typing import Any, Dict, List, Optional, Union
 
+# Third Party
 from pydantic import (
     AnyUrl,
     BaseModel,
