+++ core/tools/knowledge/file_knowledge_tool.py
