+++ core/tools/knowledge/vector_db/milvus/operations.py
@@ -1,15 +1,12 @@
-# Standard library imports
 # Standard Library
 import os
 from typing import List, Optional
 
 # Third Party
-# Third party imports
-from fastapi import FastAPI
 from pydantic import BaseModel
 
 # Local application imports
-from .CustomEmbeddings import CustomEmbeddings
+from .custom_embeddigs import CustomEmbeddings
 from .query_milvus import Milvus
 
 MILVUS_HOST = os.getenv("MILVUS_HOST", "localhost")
@@ -17,7 +14,6 @@ MILVUS_HOST = os.getenv("MILVUS_HOST", "localhost")
 model = {}
 top_re_rank = 5
 top_k_match = 10
-app = FastAPI()
 
 
 class Query(BaseModel):
@@ -26,11 +22,6 @@ class Query(BaseModel):
     topk: int = top_k_match
     rerank: bool = False
     topr: int = top_re_rank
-
-
-@app.on_event("startup")
-async def app_startup():
-    pass
 
 
 def drop_collection(collection_name: str):
@@ -56,21 +47,6 @@ def load_collection(collection_name: str) -> Milvus:
     )
 
 
-@app.post("/add_texts")
-async def add_texts_embeddings(
-    collection_name: str,
-    texts: List[str],
-    metadatas: Optional[List[dict]] = None,
-):
-    """_summary_
-
-    Args:
-        texts (List[str]): _description_
-        connlection_name (str): this should reflect user's random id + the assistant_id they created.
-    """
-    pks = add_texts(collection_name, texts, metadatas)
-
-
 def add_texts(
     collection_name: str,
     texts: List[str],
@@ -80,11 +56,6 @@ def add_texts(
     pks = c.add_texts(texts=texts, metadatas=metadatas)
     print(pks)
     return pks
-
-
-@app.delete("/delete_docs")
-async def delete_docs_api(collection_name: str, expr: str):
-    delete_docs(collection_name, expr)
 
 
 def delete_docs(collection_name: str, expr: str):
@@ -106,14 +77,3 @@ def get_top_k_biencoder_match_milvus(query: Query):
 def get_similar_match(query, biencoder_match_method: str, rerank: bool = False):
     query_biencoder_matches = get_top_k_biencoder_match_milvus(query)
     return query_biencoder_matches[: query.topr]
-
-
-@app.post("/similarity_match")
-def text_similarity_match(query: Query):
-    res = get_similar_match(query, biencoder_match_method="milvus", rerank=query.rerank)
-    return {"response": res}
-
-
-@app.get("/ping")
-def ping():
-    return {"response": "Pong!"}
