+++ pyproject.toml
@@ -1,0 +1,38 @@
+[tool.isort]
+profile = "black"
+import_heading_stdlib = "Standard Library"
+import_heading_thirdparty = "Third Party"
+import_heading_firstparty = "Local"
+known_third_party = ["openai","click", "core", "elasticsearch", "elasticsearch5", "fastapi", "importlib_resources", "flask", "gensim", "loglizer", "logparser", "matplotlib", "numpy", "pandas", "pydantic", "pytest", "regex", "requests", "scipy", "sklearn", "sompy", "sqlalchemy", "starlette", "torch", "urllib3", "yaml"]
+known_first_party = ["tests", "app"]
+multi_line_output = 3
+force_grid_wrap = 0
+use_parentheses = true
+ensure_newline_before_comments = true
+include_trailing_comma = true
+length_sort = false
+line_length = 88
+
+[tool.black]
+line-length = 88
+target-version = ['py310']
+include = '\.pyi?$'
+exclude = '''
+/(
+    \.eggs
+  | \.git
+  | \.hg
+  | \.mypy_cache
+  | \.tox
+  | \.venv
+  | _build
+  | buck-out
+  | build
+  | dist
+)/
+'''
+
+[tool.pytest.ini_options]
+pythonpath = [
+  "."
+]
